Stop rule classes persisting across tests
The CLI tests now re-require the rule files, so they become a rule class
for every test. It's a bit, sorta, hackery by mutating $LOADED_FEATURES,
and removing the constant... but it works for all tests in one process.

Primarily, if a library involves singleton-like data, then its tests
should be responsible for resetting that after every test.